// types/experience.ts
// Experience and work-related types
export interface Technology {
    name: string;
    href: string;
    icon: React.ReactNode;
}

export interface Experience {
    company: string;
    position: string;
    location: string;
    image: string;
    description: string[];
    startDate: string;
    endDate: string;
    website?: string;
    x?: string;
    linkedin?: string;
    github?: string;
    technologies: Technology[];
    isCurrent: boolean;
    isBlur?: boolean;
}